requirements.txt
asgiref==3.9.0
Django==5.2.4
djangorestframework==3.16.0
psycopg2-binary>=2.9
python-decouple==3.8
sqlparse==0.5.3
Pillow
celery
redis
requests
twilio
django-extensions
djangorestframework-simplejwt
django-cors-headers